Opti int

+++ cpp04/ex03/srcs/MateriaSource.cpp
@@ -6,7 +6,7 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/27 16:58:00 by mfeldman          #+#    #+#             */
-/*   Updated: 2024/02/02 16:27:01 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/02/04 13:28:36 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,14 @@
 MateriaSource::MateriaSource() 
 {
 	std::cout << "Type default constructor called for MateriaSource" << std::endl;
-	for (int i = 0; i < 4; ++i)
+	for (uint8_t i = 0; i < 4; ++i)
         _learned_materias[i] = NULL;
 }
 
 MateriaSource::MateriaSource(MateriaSource const &cpy) 
 {
 	std::cout << "Copy constructor called for MateriaSource" << std::endl;
-	for (int i = 0; i < 4; ++i)
+	for (uint8_t i = 0; i < 4; ++i)
     {
         if (cpy._learned_materias[i] != NULL)
             _learned_materias[i] = cpy._learned_materias[i]->clone();
@@ -34,7 +34,7 @@ MateriaSource::MateriaSource(MateriaSource const &cpy)
 MateriaSource::~MateriaSource() 
 {
 	std::cout << "Destructor called for MateriaSource" << std::endl;
-	for (int i = 0; i < 4; ++i)
+	for (uint8_t i = 0; i < 4; ++i)
         delete (_learned_materias[i]);
 }
 
@@ -43,10 +43,10 @@ MateriaSource&	MateriaSource::operator=(MateriaSource const &rhs)
     std::cout << "Copy assignment operator called for MateriaSource" << std::endl;
 	if (this != &rhs)
 	{
-        for (int i = 0; i < 4; ++i)
+        for (uint8_t i = 0; i < 4; ++i)
             delete (_learned_materias[i]);
 
-        for (int i = 0; i < 4; ++i)
+        for (uint8_t i = 0; i < 4; ++i)
         {
             if (rhs._learned_materias[i] != NULL)
                 _learned_materias[i] = rhs._learned_materias[i]->clone();
@@ -59,7 +59,7 @@ MateriaSource&	MateriaSource::operator=(MateriaSource const &rhs)
 
 void	MateriaSource::learn_materia(AMateria* m)
 {
-    for (int i = 0; i < 4; ++i)
+    for (uint8_t i = 0; i < 4; ++i)
     {
         if (_learned_materias[i] == NULL)
         {
@@ -71,7 +71,7 @@ void	MateriaSource::learn_materia(AMateria* m)
 
 AMateria*	MateriaSource::create_materia(const std::string& type)
 {
-    for (int i = 0; i < 4; ++i)
+    for (uint8_t i = 0; i < 4; ++i)
     {
         if (_learned_materias[i] != NULL && _learned_materias[i]->get_type() == type)
             return (_learned_materias[i]->clone());
